Changes after review

+++ include/tbitfield.h
@@ -12,7 +12,8 @@
 
 using namespace std;
 
-typedef unsigned int TELEM;
+//typedef unsigned int TELEM;
+typedef unsigned long long TELEM;
 
 class TBitField
 {
